+ fixed the server crashes at player login and creation. * http://bitbucket.org/maddogz/trinitycore-rollback/issue/4/server-crash-at-login + fixed SpeedChange parse packet  ( original patch by Xero ) + removed the Combat Rating System functions + fixed the receiving of spell opcode

+++ src/game/Spell.h
@@ -173,7 +173,7 @@ class SpellCastTargets
         int32 m_mapId;
         std::string m_strTarget;
 
-        uint32 m_targetMask;
+        uint16 m_targetMask;
     private:
         // objects (can be used at spell creating and after Update at casting
         Unit *m_unitTarget;
